+++ UltimateCastbars/CHANGELOGS.lua
@@ -1,6 +1,22 @@
 local _, UCB = ...
 
 UCB.CHANGELOG_TEXT = [=[
+## Version 0.9.2 - [26-02-2026]
+
+### Added
+- A control variable to modify the cancelled channel window to account for an "error" threshold (default 100ms or 0.1s)
+
+### Changed
+- Updated LibSharedMedia-3.0 to current version
+
+### Fixed
+- Update events for channel/empower spells were not behaving as intended
+- Hide changelog button on non UCB frames
+- Show/Hide preview was not working correctly in between castbar types of tabs
+- Copy profile was copying generated keys that were not in schema 
+
+---
+
 ## Version 0.9.1 - [23-02-2026]
 
 - Support for automatic Curse and Wago releases